Virtually a complete rewrite of the at_produce plugin: - Creates at-produce-all and at-produce-selected commands. - Adds the log node and redraws in the main thread after the separate thread completes. The old code was not thread safe! - Reports progress periodically using g.es_print.  This is safe because the timer executes in the main thread. - Rewrote the code using positions, modern iterators, etc. - Writes the log node as the last top-level node.
Leo build: 20140921062746

+++ leo/plugins/at_produce.py
@@ -4,115 +4,123 @@
 #@+node:ekr.20050311110307: ** << docstring >>
 ''' Executes commands in nodes whose body text starts with @produce.
 
+WARNING: trying to execute a non-existent command will hang Leo.
+
 To use, put in the body text of a node::
 
-    @produce javac -verbose Test.java
+    @produce echo hi
+    
+This plugin creates two new commands: at-produce-all and at-produce-selected.
 
-To execute, you goto Outline and look at Produce.  Choose Execute All Produce
-or Execute Tree Produce.  The Tree does the current Tree, All does the whole
-Outline.  Executing will fire javac, or whatever your using.  @produce functions
-as a directive.  After executing, a log file/node is created at the top of the
-Outline.  Any output, even error messages, should be there.
+at-produce-all scans the entire tree for body text containing @produce.
+at-produce-selected just scans the selected tree.
 
-It executes in a hierarchal manner.  Nodes that come before that contain @produce
-go first.
+Whatever follows @produce is executed as a command.
 
-I'm hoping that this orthogonal to @run nodes and anything like that.  Its not
-intended as a replacement for make or Ant, but as a simple substitute when that
-machinery is overkill.
+@produce commands are executed in the order they are found, that is, in outline order.
 
-WARNING: trying to execute a non-existent command will hang Leo.
+The at-produce commands produce a log node as the last top-level node of the outline.
+Any output, including error messages, should be there.
+
+This plugin is not intended as a replacement for make or Ant, but as a
+simple substitute when that machinery is overkill.
+
 '''
 #@-<< docstring >>
 
-#@@language python
-#@@tabwidth -4
+# 2014/09/21: EKR
+# - Creates at-produce-all and at-produce-selected commands.
+# - Adds the log node and redraws in the main thread after the separate thread completes.
 
-from __future__ import generators # To make this plugin work with Python 2.2.
-
-#@+<< imports >>
-#@+node:ekr.20040915085715: ** << imports >>
 import leo.core.leoGlobals as g
-
 import os
 import threading
 import time
 
-#@-<< imports >>
-__version__ = '.3'
-#@+<< version history >>
-#@+node:ekr.20050311110629: ** << version history >>
-#@@killcolor
-
-#@+at
-# 
-# .2 EKR:
-#     - Move all docs into docstring.
-#     - Created init function.
-#     - Use keywords dict to get c.  Removed haveseen dict.
-# 
-# .3 EKR:
-#     - Added from __future__ import generators to suppress warning in Python 2.2.
-#@-<< version history >>
-
+# Global vars.
 pr = '@' + 'produce'
 
 #@+others
+#@+node:ekr.20040915085351.7: ** addMenu (no longer used)
+def addMenu( tag, keywords ):
+    '''Produce two new entries at the end of the Outlines menu.'''
+    c = keywords.get('c')
+    if not c: return
+    mc = c.frame.menu
+    m = mc.createNewMenu ('Produce',parentName="outline",before=None)
+    c.add_command(m,
+        label = "Execute All Produce",
+        command = lambda c = c: exeProduce(c,all=True))
+    c.add_command(m,
+        label = "Execute Tree Produce",
+        command = lambda c = c: exeProduce(c,all=False ) )
+#@+node:ekr.20140920173002.17965: ** at-produce commands
+@g.command('at-produce-all')
+def produce_all_f(event):
+    c = event.get('c')
+    if c:
+        run(c,all=True)
+
+@g.command('at-produce-selected')
+def produce_selected_f(event):
+    c = event.get('c')
+    if c:
+        run(c,all=False)
 #@+node:ekr.20050311110629.1: ** init
 def init ():
-
-    ok = True # Ok for unit testing: adds menu and new directive.
-
-    if ok: 
-        g.registerHandler(('new','menu2'),addMenu)
-        g.globalDirectiveList.append('produce')
-        g.plugin_signon(__name__)
-
-    return ok
-#@+node:ekr.20040915085351.2: ** makeProduceList & allies
-def makeProduceList( c, root = True ):
-
-    pl = []
-    if root:
-        rvnode = c.rootVnode()
-        stopnode = rvnode
-    else:
-        rvnode = c.currentVnode()
-        stopnode = rvnode.next()
-
-    for z in travel(rvnode,stopnode):
-        body = z.b
-        body = body.split( '\n' )
-        body = filter(teststart, body)
-        if body:
-            map( lambda i : pl.append( i ), body )
-
-    return pl
-#@+node:ekr.20040915085351.3: *3* teststart
-def teststart( a ):
-    return a.startswith( pr )
-#@+node:ekr.20040915085351.4: *3* travel
-def travel(vn,stopnode):
-
-    while vn:
-        yield vn
-        vn = vn.threadNext()
-        if vn == stopnode:
-            vn = None
-#@+node:ekr.20040915085351.5: ** exeProduce
-def exeProduce( c,root=True):
-
-    pl = makeProduceList(c,root)
-    # Define the callback with argument bound to p1.
-    #@+others
-    #@+node:ekr.20040915085351.6: *3* runPL
-    def runPL(pl=pl):
-        # g.trace(pl)
-        f = open('produce.log', 'w+')
-        for z in pl:
+    '''Return True if the plugin has loaded successfully.'''
+    # g.registerHandler(('new','menu2'),addMenu)
+    g.globalDirectiveList.append('produce')
+    g.plugin_signon(__name__)
+    return True
+#@+node:ekr.20040915085351.5: ** run & helpers
+def run(c,all):
+    '''
+    Run all @produce nodes in a separate thread.
+    Report progress via a timer in *this* thread.
+    '''
+    aList = getList(c,all)
+    g.app.at_produce_command = None
+    def thread_target():
+        # runList must not change Leo's outline or log!
+        # runList uses c only to update c.at_produce_command.
+        runList(c,aList)
+    t = threading.Thread(target=thread_target)
+    t.setDaemon(True)
+    t.start()
+    timer = g.IdleTime(handler=None,delay=500,tag='at-produce')
+    c._at_produce_max = 20
+    c._at_produce_count = c._at_produce_max -1
+    def timer_callback(tag):
+        timer_callback_helper(c,t,timer)
+    timer.handler = timer_callback
+    timer.start()
+#@+node:ekr.20040915085351.2: *3* getList
+def getList(c,all):
+    '''
+    Return a list of all @produce lines in body texts in an outline.
+    all = True:  scan c's entire outline.
+    all = False: scan c.p and its descendants.
+    '''
+    aList = []
+    iter_ = c.all_positions if all else c.p.self_and_subtree
+    for p in iter_():
+        for line in p.b.split('\n'):
+            if line.startswith(pr):
+                aList.append(line)
+    return aList
+#@+node:ekr.20040915085351.6: *3* runList
+def runList(c,aList):
+    '''
+    Run all commands in aList (in a separate thread).
+    Do not do change Leo's outline in this thread!
+    '''
+    f = open('produce.log', 'w+')
+    try:
+        for z in aList:
             if z.startswith(pr):
-                z = z.lstrip( pr )
-                z = z.lstrip()
+                c.at_produce_command = z
+                z = z.lstrip(pr).lstrip()
                 f.write( 'produce: %s\n' % z )
                 fi, fo, fe  = os.popen3( z )
                 while 1:
@@ -127,33 +135,33 @@ def exeProduce( c,root=True):
                 fo.close()
                 fe.close() 
                 f.write('===============\n' )    
-        f.seek( 0 )
-        rv = c.rootVnode()
-        nv = rv.insertAfter()
-        c.setBodyString(nv,f.read() )
-        c.setHeadString(nv,'produce.log from %s' % time.asctime() )
+        f.seek(0)
+        s = f.read()
+    finally:
         f.close()
-        os.remove( 'produce.log' )
-    #@-others
-    t = threading.Thread(target = runPL)
-    t.setDaemon( True )
-    t.start()
-#@+node:ekr.20040915085351.7: ** addMenu
-def addMenu( tag, keywords ):
-
-    c = keywords.get('c')
-    if not c: return
-
-    mc = c.frame.menu
-
-    m = mc.createNewMenu ('Produce',parentName="outline",before=None)
-
-    c.add_command(m,
-        label = "Execute All Produce",
-        command = lambda c = c: exeProduce(c))
-
-    c.add_command(m,
-        label = "Execute Tree Produce",
-        command = lambda c = c: exeProduce(c,root=False ) )
+    
+#@+node:ekr.20140920173002.17966: *3* timer_callback_helper
+def timer_callback_helper(c,t,timer):
+    '''All drawing must be done in the main thread.'''
+    if t.isAlive():
+        c._at_produce_count += 1
+        if (c._at_produce_count % c._at_produce_max) == 0:
+            g.es_print('at-produce: %s' % c.at_produce_command)
+    else:
+        f = open('produce.log','r')
+        s = f.read()
+        f.close()
+        os.remove('produce.log' )
+        last = c.rootPosition()
+        while last and last.hasNext():
+            last.moveToNext()
+        p = last.insertAfter()
+        p.h = 'produce.log from %s' % time.asctime()
+        p.b = s
+        c.redraw()
+        timer.stop()
+        g.es_print('at-produce done')
 #@-others
+#@@language python
+#@@tabwidth -4
 #@-leo
